Label as external a test on the rst2md option
