+++ tests/external/rst2md/test_rst2md.py
@@ -3,7 +3,7 @@ import os
 import pytest
 from nbformat.v4.nbbase import new_markdown_cell, new_notebook
 
-from jupytext import read, write
+from jupytext import TextFileContentsManager, read, write
 from jupytext.cli import jupytext
 
 
@@ -43,3 +43,27 @@ def test_rst2md(tmpdir, cwd_tmpdir):
 
     # Was rst to md conversion effective?
     assert nb.cells[2].source == "$1+1$"
+
+
+@pytest.mark.requires_sphinx_gallery
+def test_rst2md_option(tmpdir):
+    tmp_py = str(tmpdir.join("notebook.py"))
+
+    # Write notebook in sphinx format
+    nb = new_notebook(
+        cells=[
+            new_markdown_cell("A short sphinx notebook"),
+            new_markdown_cell(":math:`1+1`"),
+        ]
+    )
+    write(nb, tmp_py, fmt="py:sphinx")
+
+    cm = TextFileContentsManager()
+    cm.sphinx_convert_rst2md = True
+    cm.root_dir = str(tmpdir)
+
+    nb2 = cm.get("notebook.py")["content"]
+
+    # Was rst to md conversion effective?
+    assert nb2.cells[2].source == "$1+1$"
+    assert nb2.metadata["jupytext"]["rst2md"] is False
