+++ tests/integration/contents_manager/test_contentsmanager.py
@@ -1279,30 +1279,6 @@ def test_pair_notebook_in_dotdot_folder(tmpdir):
     cm.get("scripts/notebook_name.py")
 
 
-@pytest.mark.requires_sphinx_gallery
-def test_rst2md_option(tmpdir):
-    tmp_py = str(tmpdir.join("notebook.py"))
-
-    # Write notebook in sphinx format
-    nb = new_notebook(
-        cells=[
-            new_markdown_cell("A short sphinx notebook"),
-            new_markdown_cell(":math:`1+1`"),
-        ]
-    )
-    write(nb, tmp_py, fmt="py:sphinx")
-
-    cm = jupytext.TextFileContentsManager()
-    cm.sphinx_convert_rst2md = True
-    cm.root_dir = str(tmpdir)
-
-    nb2 = cm.get("notebook.py")["content"]
-
-    # Was rst to md conversion effective?
-    assert nb2.cells[2].source == "$1+1$"
-    assert nb2.metadata["jupytext"]["rst2md"] is False
-
-
 def test_split_at_heading_option(tmpdir):
     text = """Markdown text
